[#256] Refactor: User Command / Query 로직 분리
