+++ src/main/java/com/clover/youngchat/domain/user/service/command/UserCommandService.java
@@ -1,4 +1,4 @@
-package com.clover.youngchat.domain.user.service;
+package com.clover.youngchat.domain.user.service.command;
 
 import static com.clover.youngchat.domain.user.constant.UserConstant.EMAIL_AUTHENTICATION;
 import static com.clover.youngchat.global.exception.ResultCode.ACCESS_DENY;
@@ -6,7 +6,6 @@ import static com.clover.youngchat.global.exception.ResultCode.DUPLICATED_EMAIL;
 import static com.clover.youngchat.global.exception.ResultCode.INVALID_PROFILE_IMAGE_TYPE;
 import static com.clover.youngchat.global.exception.ResultCode.MISMATCH_CONFIRM_PASSWORD;
 import static com.clover.youngchat.global.exception.ResultCode.MISMATCH_PASSWORD;
-import static com.clover.youngchat.global.exception.ResultCode.NOT_FOUND_EMAIL;
 import static com.clover.youngchat.global.exception.ResultCode.NOT_FOUND_USER;
 import static com.clover.youngchat.global.exception.ResultCode.SAME_OLD_PASSWORD;
 import static com.clover.youngchat.global.exception.ResultCode.UNAUTHORIZED_EMAIL;
@@ -19,10 +18,7 @@ import com.clover.youngchat.domain.user.dto.request.UserSignupReq;
 import com.clover.youngchat.domain.user.dto.request.UserUpdatePasswordReq;
 import com.clover.youngchat.domain.user.dto.response.UserEmailAuthCheckRes;
 import com.clover.youngchat.domain.user.dto.response.UserEmailAuthRes;
-import com.clover.youngchat.domain.user.dto.response.UserEmailCheckRes;
 import com.clover.youngchat.domain.user.dto.response.UserProfileEditRes;
-import com.clover.youngchat.domain.user.dto.response.UserProfileGetRes;
-import com.clover.youngchat.domain.user.dto.response.UserProfileSearchRes;
 import com.clover.youngchat.domain.user.dto.response.UserSignupRes;
 import com.clover.youngchat.domain.user.dto.response.UserUpdatePasswordRes;
 import com.clover.youngchat.domain.user.entity.User;
@@ -31,17 +27,18 @@ import com.clover.youngchat.global.email.EmailUtil;
 import com.clover.youngchat.global.exception.GlobalException;
 import com.clover.youngchat.global.s3.S3Util;
 import com.clover.youngchat.global.s3.S3Util.FilePath;
-import jakarta.transaction.Transactional;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 @Service
 @RequiredArgsConstructor
-public class UserService {
+@Transactional
+public class UserCommandService {
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
@@ -50,27 +47,6 @@ public class UserService {
 
     @Value("${default.image.url}")
     private String defaultProfileImageUrl;
-
-
-    public UserProfileGetRes getProfile(Long userId, User user) {
-        String email = null;
-
-        if (userId != null) {
-            user = userRepository.findById(userId)
-                .orElseThrow(() -> new GlobalException(NOT_FOUND_USER));
-        } else {
-            email = user.getEmail();
-        }
-
-        return UserProfileGetRes.to(user, email);
-    }
-
-    public UserProfileSearchRes searchProfile(String email) {
-        User user = userRepository.findByEmail(email)
-            .orElseThrow(() -> new GlobalException(NOT_FOUND_EMAIL));
-
-        return UserProfileSearchRes.to(user);
-    }
 
     public UserSignupRes signup(UserSignupReq userSignupReq) {
 
@@ -89,12 +65,9 @@ public class UserService {
         return new UserSignupRes();
     }
 
-
-    @Transactional
     public UserProfileEditRes editProfile(Long userId, UserProfileEditReq req,
         MultipartFile multipartFile,
         Long authUserId) {
-        // 수정하고자 하는 프로필이 본인의 프로필이 아닌 경우
         if (!userId.equals(authUserId)) {
             throw new GlobalException(ACCESS_DENY);
         }
@@ -104,16 +77,13 @@ public class UserService {
 
         String profileImageUrl = user.getProfileImage();
 
-        // user가 파일을 올리지 않았을 경우 프로필 변경하지 않는 것으로 간주해서 기존 프로필 그대로 유지
         if (multipartFile == null || multipartFile.isEmpty()) {
             profileImageUrl = user.getProfileImage();
         } else {
-            // user가 업로드 할 파일의 확장자가 png 인지 확인, png가 아니라면 exception 발생
             if (!Objects.equals(multipartFile.getContentType(), "image/png") &&
                 !Objects.equals(multipartFile.getContentType(), "image/jpeg")) {
                 throw new GlobalException(INVALID_PROFILE_IMAGE_TYPE);
             }
-            // user의 프로필 url이 기본 프로필과 같지 않을 경우 s3에서 삭제
             if (!profileImageUrl.equals(defaultProfileImageUrl)) {
                 s3Util.deleteFile(profileImageUrl, FilePath.PROFILE);
             }
@@ -125,7 +95,6 @@ public class UserService {
         return new UserProfileEditRes();
     }
 
-    @Transactional
     public UserUpdatePasswordRes updatePassword(Long userId, UserUpdatePasswordReq req) {
         User foundUser = findUserById(userId);
 
@@ -175,9 +144,5 @@ public class UserService {
         if (req.getPrePassword().equals(req.getNewPassword())) {
             throw new GlobalException(SAME_OLD_PASSWORD);
         }
-    }
-
-    public UserEmailCheckRes checkEmailDuplicated(final String email) {
-        return UserEmailCheckRes.to(userRepository.existsByEmail(email));
     }
 }
